Move methods which are not hermes specific which could be used by other adapters to the utils.py module
